% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{align_2_list}
\alias{align_2_list}
\title{Combine two lists into one list by aligning each item}
\usage{
align_2_list(list_a, list_b)
}
\arguments{
\item{list_a}{list_a. \code{list_a} and \code{list_b} need to have same length.}

\item{list_b}{list_b}
}
\value{
A list of same length of input list. Each item is a list of \itemize{
\item \code{a: list_a[[i]]} \item \code{b: list_b[[i]]} }
}
\description{
The generic parallel function \code{\link[=par_lapply]{par_lapply()}} can only apply a function with
single parameter to a list. Thus function with multiple parameters need to be
wrapped into a function with single list which hold all the parameters.
}
