% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{align_list}
\alias{align_list}
\title{Combine Two Lists Into One List By Aligning Each Item}
\usage{
align_list(list_a, list_b)
}
\arguments{
\item{list_a}{list_a}

\item{list_b}{list_b}
}
\value{
A list of same length of input list. Each item is a list of
\itemize{
\item \code{a: list_a[[i]]}
\item \code{b: list_b[[i]]}
}
}
\description{
\code{list_a} and \code{list_b} need to have same length.
}
