% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{align_lists}
\alias{align_lists}
\title{Combine multiple lists into one list by aligning each item}
\usage{
align_lists(...)
}
\arguments{
\item{...}{}
}
\value{
A list of same length of each input list, and each input parameter should have same length. Each item in result is a list of each \code{i}th item in input lists.
}
\description{
The generic parallel function \code{\link[=par_lapply]{par_lapply()}} can only apply a function with
single parameter to a list. Thus function with multiple parameters need to be
wrapped into a function with single parameter list.
}
\examples{
align_lists(letters[1:3], 1:3, rep_len(FALSE, length.out = 3))

}
