% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmm_data.R
\name{as_tele_list}
\alias{as_tele_list}
\title{Coerce telemetry object to list}
\usage{
as_tele_list(tele)
}
\arguments{
\item{tele}{result from \code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}}}
}
\value{
a list of telemetry objects, each named by animal name
}
\description{
\code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} will return will return single telemetry object
instead of a list when there is only one animal in data. To make code
consistent we always work with a list of telemetry objects. Either use
\code{drop = FALSE} in \code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} or apply this function over
\code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} result to ensure a proper list.
}
