% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmm_data.R
\name{assign_distance}
\alias{assign_distance}
\title{Calculate distance to median center for each animal location}
\usage{
assign_distance(animals_dt, tele_list, device_error = 10)
}
\arguments{
\item{animals_dt}{location \code{data.table} from \link{combine}}

\item{tele_list}{the telemetry obj list. Calculation need error information
from it.}

\item{device_error}{standardized device error in meter. Example: GPS: 10,
VHF: 100}
}
\value{
\code{data.table} with distance columns added. Note the original input
parameter is modified in place by reference. The name \code{assign} hint on this
nature.
}
\description{
If there are big gaps in sampling time, median center for each time group is
used. To reduce duplicate calculation, speed calculation will use some
columns created in distance calculation. Always update \code{data.table} with
\code{assign_distance} first then use \link{assign_speed}.
}
