% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_map.R
\name{base_map}
\alias{base_map}
\title{Build a leaflet base map}
\usage{
base_map(tiles_info = TILES_INFO, grid = TRUE, grid_group = GRID_GROUP)
}
\arguments{
\item{tiles_info}{A list holding tiles information. To customize it,
check and modify the default value \code{ctmmweb:::TILES_INFO}.}

\item{grid}{Add graticule by \code{leaflet::addSimpleGraticule}}

\item{grid_group}{graticule layer name, default as \code{ctmmweb:::GRID_GROUP}}
}
\value{
A leaflet widget object.
}
\description{
Build a leaflet base map
}
