% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmm_data.R
\name{calc_distance}
\alias{calc_distance}
\title{Calculate distance to median center for each animal location}
\usage{
calc_distance(animals_dt)
}
\arguments{
\item{animals_dt}{location \code{data.table} from \link{combine}}
}
\value{
\code{data.table} with distance columns added.
}
\description{
If there are big gaps in sampling time, median center for each time group is
used. To reduce duplicate calculation, speed calculation will use some
columns created in distance calculation. Always update \code{data.table} with \code{calc_distance} first then use \link{calc_speed}.
}
