% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_speed}
\alias{calculate_speed}
\title{Title}
\usage{
calculate_speed(animals_dt, device_error)
}
\arguments{
\item{animals_dt}{telemetry data in merged data.table}

\item{device_error}{device error if available}
}
\value{
data.table with speed columns added
}
\description{
Title
}
