% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmm_data.R
\name{combine}
\alias{combine}
\title{Generate combined location and info \code{data.table} from telemetry object/list}
\usage{
combine(tele_obj_list)
}
\arguments{
\item{tele_obj_list}{\code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} telemetry object/list}
}
\value{
list of - \code{data_dt}: all animals combined in one data.table - \code{info}:
animal information table
}
\description{
A \code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} telemetry list hold mutiple animal data in separate
list items, each item have the animal location data in a data frame, and
other information in various slots. This structure supports flexible S3
methods for telemetry object. However to plot multiple animals location
together with \code{ggplot2} we need to combine all location data into a single
\code{data.frame} with an animal id column.
}
\details{
This function combine any input telemetry object/List into a \code{data.table} of
location data, and another information \code{data.table} for animals. \code{data.table}
is chosen over \code{data.frame} for much better performance. This data structure
is also used in a lot of places in app, which works on any selected subset of
full data in almost all steps.
}
