% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compress_folder}
\alias{compress_folder}
\title{Compress A Folder Into Zip}
\usage{
compress_folder(folder_path, zip_name)
}
\arguments{
\item{folder_path}{The folder to be compressed}

\item{zip_name}{The name of zip}
}
\value{
The absolute path of result zip file
}
\description{
Keep the relative path structure in zip, include \code{folder_path} itself so the
zip can be extracted directly without need of creating a folder to hold
contents. The zip will be saved to one level up of \code{folder_path}.
}
