% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmmweb.R
\docType{package}
\name{ctmmweb}
\alias{ctmmweb}
\alias{ctmmweb-package}
\title{ctmmweb: A web app package for \code{ctmm}}
\description{
A shiny web app built upon \href{https://github.com/ctmm-initiative/ctmm}{ctmm}.
The main features of the app are built into a R package so you can reuse
them. The shiny app is included in the package so it's easier to install and
use.
}
\details{
\itemize{
\item \href{https://github.com/ctmm-initiative/ctmmweb}{github repo}
\item \href{https://ctmm-initiative.github.io/ctmmweb/}{reference site}
}
}
