% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_map.R
\name{get_names}
\alias{get_names}
\title{Extract vector of names from \code{data.table}}
\usage{
get_names(loc_data)
}
\arguments{
\item{loc_data}{\code{data.table} of animal locations from \code{\link[=combine]{combine()}}}
}
\value{
A character vector of animal names from \code{identity} column in
\code{loc_data}. Order is not changed.
}
\description{
Extract vector of names from \code{data.table}
}
