% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plot.R
\docType{data}
\name{BIGGER_THEME}
\alias{BIGGER_THEME}
\alias{BIGGER_KEY}
\alias{CENTER_TITLE}
\title{Override some ggplot2 default theme settings}
\format{A \code{ggplot2} function call of \code{ggplot2::theme} or \code{ggplot2::guides}}
\usage{
BIGGER_THEME

BIGGER_KEY

CENTER_TITLE
}
\description{
\itemize{
\item \code{ctmmweb:::BIGGER_THEME} increase the default ggplot2 legend, axis size
\item \code{ctmmweb:::BIGGER_KEY} increase default legend key size
\item \code{ctmmweb:::CENTER_TITLE} place title in center and bold
}
}
\keyword{datasets}
