% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_map.R
\name{heat_map}
\alias{heat_map}
\title{Build heat map from animal location data table}
\usage{
heat_map(loc_data_subset)
}
\arguments{
\item{loc_data_subset}{\code{data.table} subset of full data set of animal
locations from \link{combine}. The \code{id} column need to keep all animal names in
factor levels to maintain color consistency.}
}
\value{
A \code{Leaflet} map widget.
}
\description{
An interactive map will shown in RStudio Viewer pane when running in
interactive session. You can also further augment it with \code{leaflet}
operations, or save to a html with \code{htmlwidgets::saveWidget}.
}
