% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.R
\name{hrange_list_dt_to_formated_range_summary_dt}
\alias{hrange_list_dt_to_formated_range_summary_dt}
\title{Title}
\usage{
hrange_list_dt_to_formated_range_summary_dt(hrange_list_dt)
}
\arguments{
\item{hrange_list_dt}{a data.table holding model info and home range}
}
\value{
formated home range summary table
}
\description{
Title
}
