% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_map.R
\name{init_base_maps}
\alias{init_base_maps}
\title{Build a leaflet base map}
\usage{
init_base_maps(tiles_info = ctmmweb:::TILES_INFO)
}
\arguments{
\item{tiles_info}{A list holding tiles information. To customize it,
check and modify the default value \code{ctmmweb:::TILES_INFO}.}
}
\value{
A leaflet widget object.
}
\description{
Build a leaflet base map
}
