% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_ui_helpers.R
\name{JS.logify}
\alias{JS.logify}
\title{JS Function To Logify A \code{sliderInput}}
\usage{
JS.logify(digits = 2)
}
\arguments{
\item{digits}{digits after numerical point}
}
\value{
JS code
}
\description{
Make Shiny \code{sliderInput} \href{https://stackoverflow.com/a/39028280/3718827}{support logarithmicscale}. \code{JS.logify} create a
Javascript function in string, and \link{JS.onload} register the function to
\code{sliderInput}. Search the usage in \code{/inst/app/ui.R} for examples.
}
