% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_ui_helpers.R
\name{JS.onload}
\alias{JS.onload}
\title{Register JS Logify Function For Each \code{sliderInput}}
\usage{
JS.onload(slider_id_vec, sci = FALSE)
}
\arguments{
\item{slider_id_vec}{slider id vector}

\item{sci}{use scientific notation}
}
\value{
JS code
}
\description{
Make Shiny \code{sliderInput} \href{https://stackoverflow.com/a/39028280/3718827}{support logarithmicscale}. \code{JS.logify} create a
Javascript function in string, and \link{JS.onload} register the function to
\code{sliderInput}. Search the usage in \code{/inst/app/ui.R} for examples.
}
