% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_model_summary.R
\name{list_model_fit}
\alias{list_model_fit}
\title{Convert nested \code{ctmm::ctmm.select} result into flatten list with names}
\usage{
list_model_fit(model_fit_res)
}
\arguments{
\item{model_fit_res}{}
}
\value{
A single level list of models with names
}
\description{
Convert nested \code{ctmm::ctmm.select} result into flatten list with names
}
