% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.R
\name{merge_animals}
\alias{merge_animals}
\title{Generate merged data.table and info table from telemetry object/list}
\usage{
merge_animals(tele_objs)
}
\arguments{
\item{tele_objs}{telemetry object/list}
}
\value{
list of \code{data}: all animals merged in one data.table, \code{info}: animal
information table
}
\description{
Generate merged data.table and info table from telemetry object/list
}
\examples{
merge_animals(buffalo)
}
