% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmm_data.R
\name{merge_tele}
\alias{merge_tele}
\title{Generate merged location and info \code{data.table} from telemetry object/list}
\usage{
merge_tele(tele_obj_list)
}
\arguments{
\item{tele_obj_list}{telemetry object/list}
}
\value{
list of
\itemize{
\item \code{data}: all animals merged in one data.table
\item \code{info}: animal information table
}
}
\description{
A Telemetry list hold mutiple animal data in separate list items, each item
have the animal location data in a data frame, and other information in
various slots. This structure supports flexible S3 methods for telemetry
object. However to plot multiple animals location together with \code{ggplot2} we
need to merge all location data into a single \code{data.frame} with an animal id
column.
}
\details{
This function merge any input telemetry object/List into a \code{data.table} of
location data, and another information \code{data.table} for animals. \code{data.table}
is chosen over \code{data.frame} for much better performance. This data structure
is also used in a lot of places in app, which works on any selected subset of
full data in almost all steps.
}
