% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_vario_hrange_occurrence.R
\name{multi_title}
\alias{multi_title}
\alias{multi_color_title}
\alias{multi_color_multi_line_title}
\title{Multiple Colored Plot Title}
\usage{
multi_title(...)

multi_color_title(color_vec, text_vec, sep = ", ")

multi_color_multi_line_title(color_vec, text_vec, sep = "\\n\\n")
}
\arguments{
\item{...}{Color and text interwoven in order. See example. Note some spaces
need to be placed around text.}

\item{color_vec}{Vector of colors. Should have same length with \code{text_vec}}

\item{text_vec}{Vector of text segments. Will have the cooresponding color in
\code{color_vec}.}

\item{sep}{Separator to be placed among the text segments}
}
\description{
Make multiple colored base plot title.
}
\section{Functions}{
\itemize{
\item \code{multi_title}: The original function taken from
\href{http://r.789695.n4.nabble.com/title-words-in-different-colors-td878698.html#a878700}{mailing list by Barry Rowlingson}.
The only change is the \code{NULL} return value was made invisible.

\item \code{multi_color_title}: Instead of taking interwoven color and text input,
color vector and text vector are used.

\item \code{multi_color_multi_line_title}: Place each text segment in a new line.
}}

\examples{
plot(1, 2)
multi_title(color="red", "Traffic",
           color="orange", " light ",
           color="green", "signal")
}
