% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_model_summary.R
\name{overlap_2d_to_1d}
\alias{overlap_2d_to_1d}
\title{Convert overlap result 2d table to combination rows}
\usage{
overlap_2d_to_1d(overlap_matrix_dt)
}
\arguments{
\item{overlap_matrix_dt}{2d data.table from \link{overlap_matrix_to_dt} applied on
overlap matrix}
}
\value{
a data.table with overlap pair combinations in rows
}
\description{
Convert overlap result 2d table to combination rows
}
