% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{par_fit_tele}
\alias{par_fit_tele}
\title{Parallel fit models on telemetry list}
\usage{
par_fit_tele(tele_list, reserved_cores = 0, fallback = FALSE)
}
\arguments{
\item{tele_list}{telemetry list}

\item{reserved_cores}{reserve some cores so that not all cores are used.
Check your platform's core count with
\code{parallel::detectCores(logical = FALSE)}. \code{?parallel::detectCores}
has more details about physical/logical cores in different platform.}

\item{fallback}{Disable parallel and use regular \code{lapply}.}
}
\value{
list of model fitting results on each telemetry object by \code{ctmm::ctmm.select}, named by
telemetry object names
}
\description{
Parallel fit models on telemetry list
}
