% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{par_hrange_each}
\alias{par_hrange_each}
\title{Parallel calculate home range separately}
\usage{
par_hrange_each(
  tele_list,
  model_list,
  weight_list,
  cores = NULL,
  parallel = TRUE
)
}
\arguments{
\item{weight_list}{List of True/False in same order of tele_list. Used for
optimal weighting parameter in \code{\link[ctmm:akde]{ctmm::akde()}}.}

\item{cores}{the core count to be used for cluster. Could be a positive
integer or
\itemize{
\item Default \code{NULL} value will indicate to use a heuristic value based on detected cores, which is roughly \code{min(input_size, physical_cores_count * n)},
\code{n} being 2 for windows, 4 for Mac/Linux. See \code{\link[parallel:detectCores]{parallel::detectCores()}} for
more information on physical/logical cores in different platforms.
\item A negative value like \code{-2} will use \verb{all available cores - 2}, so that 2
cores are reserved for user's other tasks.
}}

\item{parallel}{Use regular \code{\link[=lapply]{lapply()}} when FALSE. You may notice more console
messages in this mode because the console messages in parallel mode are
lost by default as they happened in other threads.}
}
\value{
List of home ranges
}
\description{
Sometimes you may want to calculate home range individually, because
calculate them in same grid could take significant memory when individuals
are spread out.
}
