% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{par_occur}
\alias{par_occur}
\title{Parallel Calculate Occurrence From Telemetry And Model List}
\usage{
par_occur(tele_model_list, fallback = FALSE)
}
\arguments{
\item{tele_model_list}{Aligned list of telemetry list and model list}

\item{fallback}{Use regular \code{lapply}. This is used to test if parallel caused
problems.}
}
\value{
occurrence results list
}
\description{
Parallel Calculate Occurrence From Telemetry And Model List
}
\examples{
\dontrun{
  para_ll_ud(align_list(buffalo, models_list))
}
}
