% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.R
\name{para_ll_fit}
\alias{para_ll_fit}
\title{Title}
\usage{
para_ll_fit(tele_list)
}
\arguments{
\item{tele_list}{telemetry list}
}
\value{
list of model fitting results on each telemetry object
}
\description{
Title
}
\examples{
para_ll_fit(buffalo)
}
