% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{para_ll_fit_tele_guess}
\alias{para_ll_fit_tele_guess}
\title{Parallel Fit Models For List Of Telemetry List And Guess List}
\usage{
para_ll_fit_tele_guess(tele_guess_list)
}
\arguments{
\item{tele_guess_list}{aligned list of telemetry list and guess list}
}
\value{
list of model fitting results on each telemetry object
}
\description{
Parallel Fit Models For List Of Telemetry List And Guess List
}
