% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{pick}
\alias{pick}
\alias{pick.telemetry}
\alias{pick.list}
\title{Pick subset from telemetry object/list}
\usage{
pick(object, m)

\method{pick}{telemetry}(tele, m)

\method{pick}{list}(tele_list, m)
}
\arguments{
\item{m}{subset size}

\item{tele}{telemetry object}

\item{tele_list}{telemetry list}
}
\value{
telemetry object with less data points

telemetry list of subsets
}
\description{
A dataset subset can have models fitted much quicker. This is used to reduce
waiting time in developing code that involved time consuming modeling
processes. After code is tested and stablized, full size dataset can be used.
}
\details{
\code{m} even spaced points are taken from each object.
}
\section{Methods (by class)}{
\itemize{
\item \code{telemetry}: subset from telemetry object

\item \code{list}: list of subset from each telemetry object
}}

