% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.R
\name{pick_m_tele}
\alias{pick_m_tele}
\title{Title}
\usage{
pick_m_tele(tele, m)
}
\arguments{
\item{tele}{telemetry object}

\item{m}{sample size}
}
\value{
sampled telemetry object
}
\description{
Title
}
\examples{
pick_m_tele(buffalo[[1]], 100)
}
