% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.R
\name{pick_m_tele_list}
\alias{pick_m_tele_list}
\title{Title}
\usage{
pick_m_tele_list(tele_list, m)
}
\arguments{
\item{tele_list}{telemetry list}

\item{m}{sample size}
}
\value{
sampled telemetry list
}
\description{
Title
}
\examples{
pick_m_tele_list(buffalo, 100)
}
