% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plot.R
\name{plot_loc}
\alias{plot_loc}
\title{Plot animal locations subset with background}
\usage{
plot_loc(selected_dt, dt, point_size = 0.1, overlay_all = TRUE)
}
\arguments{
\item{selected_dt}{A subset of animals. A \code{data.table} subset taken from
\code{dt}.}

\item{dt}{The full data set of all animals location in merged \code{data.table}
format. Usually came from \link{merge_animals}.}

\item{point_size}{The size of point in plot. Denser plot may need smaller
point size.}

\item{overlay_all}{Draw full data set as background}
}
\value{
A ggplot object.
}
\description{
Draw a scatter plot of some animals' location with \code{ggplot2} (like plot 2
Overview in app). To color each individual consistently and add all animals
as background, the full data set of animals is also needed.
}
\examples{
library(ctmm)
library(ctmmweb)
data("buffalo")
dt <- merge_animals(buffalo)$data
selected_dt <- dt[identity \%in\% c("Gabs", "Queen")]
plot_loc(selected_dt, dt)
# you can take the ggplot2 object to further customize it
plot_loc(selected_dt, dt) +
  ggplot2::ggtitle("Locations of Buffalos") +
  # override the default left alignment of title and make it bigger
  ctmmweb:::CENTER_TITLE
# or export plot as png
g <- plot_loc(selected_dt, dt)
ggplot2::ggsave("test.png", g)
}
