% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plot.R
\name{plot_time}
\alias{plot_time}
\title{Plot animal sampling time histogram in facet}
\usage{
plot_time(dt_subset)
}
\arguments{
\item{dt_subset}{A subset of animals. A \code{data.table} subset taken from full
data set of all animals location. Note the \code{id} column is factor with all
animals identity as levels. So this subset still have all identities
information, which is needed to color each animal consistently.}
}
\value{
A ggplot object.
}
\description{
Plot animal sampling time histogram in facet
}
