% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_vario_hrange_occurrence.R
\name{plot_ud}
\alias{plot_ud}
\title{Plot a group of home ranges or occurrences}
\usage{
plot_ud(UD_list, level_vec = 0.95, columns = 2, tele_list = NULL)
}
\arguments{
\item{UD_list}{\code{ctmm} \code{UD} object list, which can be home range from
\code{ctmm::akde} or occurrences from \code{ctmm::occurrence}. The names of list are
needed for title of figures.}

\item{level_vec}{The vector of \code{level.UD} in \code{ctmm::plot.telemetry}. To be
consistent with \code{ctmm} they are values 0 ~ 1 (for example 0.95). Note the
app UI take values 0 ~ 100 (for example 95) for easier input.}

\item{columns}{The columns of the group plot layout.}

\item{tele_list}{Overlay animal locations when provided. This should only be
used for home range since it can interfere with occurrence plot.}
}
\description{
This is a wrapper over \code{ctmm::plot.telemetry} to make group plot easier.
}
