% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_vario_hrange_occurrence.R
\name{plot_vario}
\alias{plot_vario}
\title{Plot a group of empirical or modeled variograms.}
\usage{
plot_vario(
  vario_list,
  model_list = NULL,
  title_vec = NULL,
  fraction = 0.5,
  relative_zoom = TRUE,
  cex = 0.65,
  model_color = "blue",
  columns = 2
)
}
\arguments{
\item{vario_list}{list of \code{\link[ctmm:variogram]{ctmm::variogram()}}. The names of list may be
needed for figure titles.}

\item{model_list}{\code{CTMM} models list. The \code{names} of list may be needed for
figure titles. Draw modeled variogram if provided. The models list should
match \code{vario_list} in length and animal, so that \code{i}th model is for \code{i}th
variogram.
Possible values:
\itemize{
\item It can be skipped to use default value \code{NULL}, and only empirical
variograms are drawn. \code{vario_list} names will be used as figure title.
Better use named parameters for others in this case.
\item list of guesstimated models from \code{\link[ctmm:variogram.fit]{ctmm::ctmm.guess()}} on telemetry data,
overlay guesstimate variogram.
\item list of fitted models from \code{\link[ctmm:ctmm.fit]{ctmm::ctmm.select()}} on \code{vario_list}, overlay
modeled variogram.
\item It can be a list of list, with each item provide list of \code{CTMM} objects
that to be plotted in each variogram. \code{model_color} can be a vector
to specify color for each model.
}}

\item{title_vec}{vector of figure titles. If not provided, names of
\code{vario_list} or \code{model_list} will be used.}

\item{fraction}{Fraction of time-lag range, 0 ~ 1.}

\item{relative_zoom}{\itemize{
\item Relative mode zoom every plot by fraction of their own Time-lag range. The
X, Y axes are not synced.
\item Absolute mode operate on the max Time-lag range individual in group, and
all others scaled with same X, Y axes for easier comparison.
}}

\item{cex}{The magnification factor of plot text and symbols. See \code{\link[=par]{par()}}.
You may need a smaller value if figure margins is too large}

\item{model_color}{The color of model variogram. It'll be same for every
plot.}

\item{columns}{The columns of the group plot layout.}
}
\description{
This is a wrapper over \code{\link[ctmm:plot.variogram]{ctmm::plot.variogram()}} to make group plot easier.
}
