% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_map.R
\name{point_map}
\alias{point_map}
\title{Build point map from animal location data table}
\usage{
point_map(selected_dt, dt)
}
\arguments{
\item{selected_dt}{\code{data.table} subset of \code{dt}.}

\item{dt}{Full data set of \code{data.table} of animal locations from \link{merge_tele}.}
}
\value{
A \code{Leaflet} map widget.
}
\description{
The full data set is also needed to maintain color consistency.
}
