% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_map.R
\name{range_map}
\alias{range_map}
\title{Build home range map}
\usage{
range_map(hrange_list, hr_levels, hr_color_vec)
}
\arguments{
\item{hrange_list}{list of home range UD object. The names of list will be
used as layer names, usually they are model names.}

\item{hr_levels}{The vector of \code{level.UD} in \code{ctmm::plot.telemetry}. To be
consistent with \code{ctmm} they are values 0 ~ 1 (for example 0.95). Note the
app UI take values 0 ~ 100 (for example 95) for easier input.}

\item{hr_color_vec}{Vector of color names to be used for each home range. The
length of \code{hr_color_vec} should match length of \code{hrange_list}.}
}
\value{
A \code{Leaflet} map widget.
}
\description{
An interactive map will shown in RStudio Viewer pane when running in
interactive session. You can also further augment it with \code{leaflet}
functions, or save to a html with \code{htmlwidgets::saveWidget}.
}
