% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{resampling.list}
\alias{resampling.list}
\title{Resample each telemetry object in list}
\usage{
\method{resampling}{list}(tele_list, m)
}
\arguments{
\item{tele_list}{telemetry list}

\item{m}{sample size}
}
\value{
resampled telemetry list
}
\description{
Resample each object with \link{resampling}.
}
