% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{resampling.telemetry}
\alias{resampling.telemetry}
\title{Resample telemetry object}
\usage{
\method{resampling}{telemetry}(tele, m)
}
\arguments{
\item{tele}{telemetry object}

\item{m}{sample size. \code{m} even spaced points are taken from data.}
}
\value{
resampled telemetry object
}
\description{
Like image resampling, a resampled dataset can have models
fitted much quicker. This is used to reduce waiting time in developing code
that involved time consuming modeling processes. After code is tested and
stablized, full size dataset can be used.
}
