% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plot.R
\name{safe_print}
\alias{safe_print}
\title{Print/Save plot in Shiny safely}
\usage{
safe_print(expr)
}
\arguments{
\item{expr}{An expression that print/save plots.}
}
\description{
If there is open graphic device in R session (for example, just made some
plots), saving plot to file through \code{ggplot2::ggsave} or other print methods
inside Shiny may mess up the graphic device of Shiny, and all plots will go
to the RStudio plot window or R plot window.
\href{https://stackoverflow.com/questions/47699956/ggplot-in-shiny-app-go-to-rstudio-plot-window/}{Example}
}
\details{
This function simply save current graphic device before printing/saving and
restore it later.
}
\examples{
\dontrun{
  safe_print(ggsave("plot.png", g))
}
}
