% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sample_tele}
\alias{sample_tele}
\title{Sample From Telemetry Object}
\usage{
sample_tele(tele, m)
}
\arguments{
\item{tele}{telemetry object}

\item{m}{sample size}
}
\value{
sampled telemetry object
}
\description{
Take even spaced \code{m} points. Rely on ctmm S3 method to treat telemetry object
as a data.frame, thus ctmm need to be imported in NAMESPACE.
}
\examples{
sample_tele(buffalo[[1]], 100)
}
