% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{sample_tele_list}
\alias{sample_tele_list}
\title{Sample Each Telemetry Object In List}
\usage{
sample_tele_list(tele_list, m)
}
\arguments{
\item{tele_list}{telemetry list}

\item{m}{sample size}
}
\value{
sampled telemetry list
}
\description{
Sample each object with \link{sample_tele}.
}
\examples{
library(ctmm)
data(buffalo)
sample_tele_list(buffalo, 100)
}
