% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_model_summary.R
\name{summary_model_fit}
\alias{summary_model_fit}
\title{Generate formated model summary table from model fit results}
\usage{
summary_model_fit(model_fit_res)
}
\arguments{
\item{model_fit_res}{list of applying \code{ctmm::ctmm.select} on telemetry objects}
}
\value{
A \code{data.table} of model summary
}
\description{
Generate formated model summary table from model fit results
}
