% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tele_list_info}
\alias{tele_list_info}
\title{Get Information Table For Telemetry Object Or List}
\usage{
tele_list_info(tele_obj_list)
}
\arguments{
\item{tele_obj_list}{telemetry object or list}
}
\value{
An information \code{data.table} for input
}
\description{
Get Information Table For Telemetry Object Or List
}
