% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{WIN_INIT_ctmm}
\alias{WIN_INIT_ctmm}
\title{Expression to be initialized in Windows for \code{ctmm} related parallel
operations}
\format{An object of class \code{expression} of length 1.}
\usage{
WIN_INIT_ctmm
}
\description{
Parallel cluter in Windows is a socket cluster, which need to initialize each
session manually. For ctmm related parallel operations, \code{ctmm} package need
to be loaded. Instead of \code{library(ctmm)}, the expression of
\code{requireNamespace("ctmm", quietly = TRUE)} is more appropriate inside a
package.
}
\keyword{datasets}
