% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_files.R
\name{zip_folder}
\alias{zip_folder}
\title{Compress a folder into zip}
\usage{
zip_folder(folder_path, zip_name)
}
\arguments{
\item{folder_path}{The folder to be compressed}

\item{zip_name}{The name of zip}
}
\value{
The absolute path of result zip file
}
\description{
\code{\link[zip:zip]{zip::zip()}} is used so this is platform independent. It will keep the
relative path structure in zip, include \code{folder_path} itself so the zip can
be extracted directly without need of creating a folder to hold contents. The
zip file will be saved to the parent folder of \code{folder_path}.
}
