% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_map.R
\name{add_control}
\alias{add_control}
\title{Add layer control for leaflet map}
\usage{
add_control(leaf, layer_vec)
}
\arguments{
\item{leaf}{leaflet map widget object.}

\item{layer_vec}{character vector of user data layers. For example animal
names, model names etc.}
}
\description{
Add layer control for leaflet map
}
