% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_map.R
\name{add_control}
\alias{add_control}
\title{Add layer control for leaflet map}
\usage{
add_control(leaf, layer_vec, tiles_info = TILES_INFO)
}
\arguments{
\item{leaf}{leaflet map widget object.}

\item{layer_vec}{character vector of user data layers. For example animal
names, model names etc.}

\item{tiles_info}{A list holding tiles information, default to be
\code{ctmmweb:::TILES_INFO}. To customize it, use \code{ctmmweb:::TILES_INFO} as a
template.}
}
\description{
Add layer control for leaflet map
}
