% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.R
\name{align_list}
\alias{align_list}
\title{Title}
\usage{
align_list(list_a, list_b)
}
\arguments{
\item{list_a}{list_a}

\item{list_b}{list_b}
}
\value{
A list of list, each sublist have two item from list_a and list_b
}
\description{
Title
}
