% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Start ctmm web app}
\usage{
app(shiny_app_data = NULL)
}
\arguments{
\item{data}{Data to be loaded with app. It could be anything can be taken by \code{ctmm::as.telemetry}:
\itemize{
\item Path to MoveBank csv file
\item MoveBank \code{data.frame} object
\item \code{Move} object
\item \code{ctmm} `telemetry`` object/list.
\item Default \code{NULL} will not load any data.
}}
}
\description{
Start ctmm web app
}
