% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Start ctmm web app}
\usage{
app(data = NULL)
}
\arguments{
\item{data}{Data to be loaded with app. Could be path to csv file, .rds file; \code{ctmm} telemetry object/list. Default \code{NULL} will not load any data.}
}
\description{
Start ctmm web app
}
