% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmm_data.R
\name{as_tele_list}
\alias{as_tele_list}
\title{Coerce telemetry object to list}
\usage{
as_tele_list(tele)
}
\arguments{
\item{tele}{result from ctmm::as.telemetry, either a single telemetry
object(when data only have one individual), or a list of telemetry
objects(with multiple individuals)}
}
\value{
a list of telemetry objects, each named by animal name
}
\description{
Coerce telemetry object to list
}
