% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_map.R
\name{base_map}
\alias{base_map}
\title{Build a leaflet base map}
\usage{
base_map(tiles_info = TILES_INFO, grid = TRUE)
}
\arguments{
\item{tiles_info}{A list holding tiles information, default to be
\code{ctmmweb:::TILES_INFO}. To customize it, use \code{ctmmweb:::TILES_INFO} as a
template.}

\item{grid}{Add graticule by \code{\link[leaflet:addSimpleGraticule]{leaflet::addSimpleGraticule()}}}
}
\value{
A leaflet widget object.
}
\description{
Build a leaflet base map
}
