% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_hrange_list_dt}
\alias{build_hrange_list_dt}
\title{Build Home Range list table}
\usage{
build_hrange_list_dt(selected_model_names_dt, selected_hrange_list)
}
\arguments{
\item{selected_hrange_list}{home range list}

\item{selected_dt}{model names \code{data.table}}
}
\value{
a data.table holding model info and home range
}
\description{
The table structure is similar to model list table, with model information
from model summary table, and home range objects as list column
}
