% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Title}
\usage{
calculate_distance(animals_dt)
}
\arguments{
\item{animals_dt}{telemetry data in merged data.table}
}
\value{
data.table with distance columns added
}
\description{
Title
}
