% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmm_data.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate distance to median center for each animal location}
\usage{
calculate_distance(animals_dt)
}
\arguments{
\item{animals_dt}{telemetry data in merged data.table. See \code{\link[=merge_tele]{merge_tele()}}}
}
\value{
data.table with distance columns added. Note the input parameter is
modified in place.
}
\description{
If there are big gaps in sampling time, median center for each time group is
used. To reduce duplicate calculation, speed calculation will use some
columns created in distance calculation. Always \code{calculate_distance} first
then \code{\link[=calculate_speed]{calculate_speed()}}.
}
