% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmm_data.R
\name{collect}
\alias{collect}
\title{Collect location and info \code{data.table} from telemetry list}
\usage{
collect(tele_obj_list)
}
\arguments{
\item{tele_list}{\code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} telemetry list. Use \code{\link[=as_tele_list]{as_tele_list()}}
over \code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} to ensure a proper list.}
}
\value{
list of - \code{data_dt}: all animals collected in one data.table -
\code{info}: animal information table
}
\description{
A \code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} telemetry list hold mutiple animal data in separate
list items, each item have the animal location data in a data frame, and
other information in various slots. This structure supports flexible S3
methods for telemetry object. However to plot multiple animals location
together with \code{ggplot2} we need to collect all location data as a single
\code{data.frame} with an animal id column.
}
\details{
This function convert any input telemetry List into a list of 1. \code{data.table}
of location data, and 2. animal information \code{data.table}. \code{data.table} is
chosen over \code{data.frame} for much better performance. This data structure is
also used in a lot of places in app, which works on any selected subset of
full data in almost all steps.
}
