% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compress_folder}
\alias{compress_folder}
\title{Title}
\usage{
compress_folder(folder_path, zip_name)
}
\arguments{
\item{folder_path}{The folder to be compressed}

\item{zip_name}{The name of zip}
}
\value{
The absolute path of result zip file
}
\description{
Title
}
