% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmmweb.R
\docType{package}
\name{ctmmweb}
\alias{ctmmweb}
\alias{ctmmweb-package}
\title{ctmmweb: A web app package for \code{ctmm}}
\description{
It's a shiny web app built upon
\href{https://github.com/ctmm-initiative/ctmm}{ctmm}. This is also a R package so
you can reuse some functionalities in your own R analysis.
}
\details{
\itemize{
\item \href{https://github.com/ctmm-initiative/ctmm-webapp}{github repo}
\item \href{https://ctmm-initiative.github.io/ctmm-webapp/}{reference site}
}
}
