% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_model_summary.R
\name{flatten_models}
\alias{flatten_models}
\title{Convert nested \link{par_fit_tele} result into flatten list with model names}
\usage{
flatten_models(model_fit_res)
}
\arguments{
\item{model_fit_res}{}
}
\value{
A single level list of models with names
}
\description{
Convert nested \link{par_fit_tele} result into flatten list with model names
}
