% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_map.R
\name{get_names}
\alias{get_names}
\title{Extract vector of names from \code{data.table}}
\usage{
get_names(dt)
}
\arguments{
\item{dt}{\code{data.table} of animal locations from \link{merge_tele}}
}
\value{
A character vector of animal names from \code{identity} column in \code{dt}.
Order is not changed.
}
\description{
Extract vector of names from \code{data.table}
}
