% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_map.R
\name{heat_map}
\alias{heat_map}
\title{Generate heat map from animal location data table}
\usage{
heat_map(selected_dt)
}
\arguments{
\item{selected_dt}{\code{data.table} (subset) of animal locations from \link{merge_tele}}
}
\value{
A \code{Leaflet} map widget.
}
\description{
An interactive map will shown in RStudio Viewer pane when running in
interactive session. You can also further augment it with \code{leaflet}
operations, or save to a html with \code{htmlwidgets::saveWidget}.
}
