% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.R
\name{JS.logify}
\alias{JS.logify}
\title{Title}
\usage{
JS.logify(digits = 2)
}
\arguments{
\item{digits}{digits after numerical point}
}
\value{
JS function code
}
\description{
Check relevant source code in ui.R for usage examples
}
