% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{JS.logify}
\alias{JS.logify}
\title{JS Function To Logify A \code{sliderInput}}
\usage{
JS.logify(digits = 2)
}
\arguments{
\item{digits}{digits after numerical point}
}
\value{
JS function code
}
\description{
To be added in UI code of Shiny. The server code of Shiny may also use it
when building UI dynamically. Search the usage in \code{/inst/app/ui.R} for
examples.
}
