% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{JS.onload}
\alias{JS.onload}
\title{Register JS Logify Function For Each \code{sliderInput}}
\usage{
JS.onload(slider_id_vec, sci = FALSE)
}
\arguments{
\item{slider_id_vec}{slider id vector}

\item{sci}{use scientific notation}
}
\value{
JS functions code
}
\description{
Register JS Logify Function For Each \code{sliderInput}
}
