% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.R
\name{model_list_dt_to_formated_model_summary_dt}
\alias{model_list_dt_to_formated_model_summary_dt}
\title{Title}
\usage{
model_list_dt_to_formated_model_summary_dt(models_dt)
}
\arguments{
\item{models_dt}{model list dt}
}
\value{
formated model summary table
}
\description{
Title
}
