% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_model_summary.R
\name{overlap_matrix_to_dt}
\alias{overlap_matrix_to_dt}
\title{Convert overlap result matrix to data.table}
\usage{
overlap_matrix_to_dt(mat_3d, clear_half = TRUE)
}
\arguments{
\item{mat_3d}{the 3d matrix result from \code{\link[ctmm:overlap]{ctmm::overlap()}}}

\item{clear_half}{whether to clear the lower triangular part, including the
diagonal}
}
\value{
a \code{data.table} of overlap results
}
\description{
Convert overlap result matrix to data.table
}
