% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{par_fit_models}
\alias{par_fit_models}
\title{Parallel fitting model on telemetry list}
\usage{
par_fit_models(tele_list, reserved_cores = 0, parallel = TRUE)
}
\arguments{
\item{tele_list}{telemetry list}

\item{reserved_cores}{reserve some cores so that not all cores are used.
Check your platform's core count with
\code{parallel::detectCores(logical = FALSE)}. \code{?parallel::detectCores}
has more details about physical/logical cores in different platforms.}

\item{parallel}{Use regular \code{lapply} when FALSE}
}
\value{
list of models named by animal names
}
\description{
\code{ctmm::ctmm.fit} fit model on telemetry object. This can be parallelized on
each object in list.
}
