% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{par_fit_models}
\alias{par_fit_models}
\title{Parallel fitting model on telemetry list}
\usage{
par_fit_models(tele_list, cores = NULL, parallel = TRUE)
}
\arguments{
\item{tele_list}{telemetry list}

\item{cores}{the cores count to be used for cluster. Could be a positive
integer or
\itemize{
\item Default \code{NULL} value will indicate to use a heuristic value based on detected cores, which is roughly \code{min(input_size, physical_cores_count * n)},
\code{n} being 2 for windows, 4 for Mac/Linux. See \code{?parallel::detectCores} for
more information on physical/logical cores in different platforms.
\item A negative value like \code{-2} will use \code{all available cores - 2}, so that 2
cores are reserved for user's other tasks.
}}

\item{parallel}{Use regular \code{lapply} when FALSE. You may notice more console
messages in this mode because the console messages in parallel mode are
lost by default as they happened in other threads.}
}
\value{
list of models named by animal names
}
\description{
\code{ctmm::ctmm.fit} fit model on telemetry object. This function run
\code{ctmm::ctmm.fit} on each object of list on parallel.
}
