% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{par_fit_tele}
\alias{par_fit_tele}
\title{Parallel fit models on telemetry list}
\usage{
par_fit_tele(tele_list, fallback = FALSE)
}
\arguments{
\item{tele_list}{telemetry list}

\item{fallback}{Use regular \code{lapply}. This is used to test if parallel caused
problems.}
}
\value{
list of model fitting results on each telemetry object, named by
telemetry object names
}
\description{
Parallel fit models on telemetry list
}
\examples{
\dontrun{
  library(ctmm)
  data(buffalo)
  par_fit_tele(sample_tele_list(buffalo, 100))
}
}
