% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{par_occur}
\alias{par_occur}
\title{Parallel calculate occurrence from telemetry and model list}
\usage{
par_occur(tele_list, model_list, fallback = FALSE)
}
\arguments{
\item{tele_list}{\code{ctmm} \code{telemetry} list}

\item{model_list}{Corresponding \code{ctmm} model list for \code{tele_list}}

\item{fallback}{Use regular \code{lapply}. This is used to test if parallel caused
problems.}
}
\value{
occurrence results list
}
\description{
Parallel calculate occurrence from telemetry and model list
}
