% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{par_occur}
\alias{par_occur}
\title{Parallel calculate occurrence from telemetry and model list}
\usage{
par_occur(tele_list, model_list, cores = NULL, parallel = TRUE)
}
\arguments{
\item{tele_list}{\code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} telemetry list}

\item{model_list}{Corresponding \code{ctmm} model list for \code{tele_list}}

\item{cores}{the core count to be used for cluster. Could be a positive
integer or
\itemize{
\item Default \code{NULL} value will indicate to use a heuristic value based on detected cores, which is roughly \code{min(input_size, physical_cores_count * n)},
\code{n} being 2 for windows, 4 for Mac/Linux. See \code{\link[parallel:detectCores]{parallel::detectCores()}} for
more information on physical/logical cores in different platforms.
\item A negative value like \code{-2} will use \verb{all available cores - 2}, so that 2
cores are reserved for user's other tasks.
}}

\item{parallel}{Use regular \code{\link[=lapply]{lapply()}} when FALSE. You may notice more console
messages in this mode because the console messages in parallel mode are
lost by default as they happened in other threads.}
}
\value{
occurrence results list
}
\description{
Parallel calculate occurrence from telemetry and model list
}
