% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{par_try_models}
\alias{par_try_models}
\alias{par_fit_models}
\title{Parallel fitting models on telemetry list}
\usage{
par_try_models(tele_list, cores = NULL, parallel = TRUE)

par_fit_models(tele_list, cores = NULL, parallel = TRUE)
}
\arguments{
\item{tele_list}{\code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} telemetry list}

\item{cores}{the core count to be used for cluster. Could be a positive
integer or
\itemize{
\item Default \code{NULL} value will indicate to use a heuristic value based on detected cores, which is roughly \code{min(input_size, physical_cores_count * n)},
\code{n} being 2 for windows, 4 for Mac/Linux. See \code{\link[parallel:detectCores]{parallel::detectCores()}} for
more information on physical/logical cores in different platforms.
\item A negative value like \code{-2} will use \verb{all available cores - 2}, so that 2
cores are reserved for user's other tasks.
}}

\item{parallel}{Use regular \code{\link[=lapply]{lapply()}} when FALSE. You may notice more console
messages in this mode because the console messages in parallel mode are
lost by default as they happened in other threads.}
}
\value{
\code{par_try_models}: list of items named by animal names, each item hold
the attempted models as sub items with model type as name.

\code{par_fit_models}: list of models named by animal names.
}
\description{
par_try_models run \code{\link[ctmm:ctmm.select]{ctmm::ctmm.select()}} on each object of list on parallel.

par_fit_models run \code{\link[ctmm:ctmm.fit]{ctmm::ctmm.fit()}} on each object of list on parallel.
}
\section{Functions}{
\itemize{
\item \code{par_fit_models}: 
}}

