% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{par_try_models}
\alias{par_try_models}
\title{Parallel selecting models on telemetry list}
\usage{
par_try_models(tele_list, reserved_cores = 0, parallel = TRUE)
}
\arguments{
\item{tele_list}{telemetry list}

\item{reserved_cores}{reserve some cores so that not all cores are used.
Check your platform's core count with
\code{parallel::detectCores(logical = FALSE)}. \code{?parallel::detectCores}
has more details about physical/logical cores in different platforms.}

\item{parallel}{Use regular \code{lapply} when FALSE}
}
\value{
list of items named by animal names, each item hold the attempted
models as sub items with model type as name
}
\description{
\code{ctmm::ctmm.select} fit multiple models on telemetry object and output list
of all attempted models. This function run \code{ctmm::ctmm.select} on each object
of list on parallel.
}
