% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.R
\name{para_ll}
\alias{para_ll}
\title{Parallel Apply Function To List In All Platforms}
\usage{
para_ll(ll, fun)
}
\arguments{
\item{ll}{A list to apply fun}

\item{fun}{Function to be applied}
}
\value{
List of applied results
}
\description{
Parallel Apply Function To List In All Platforms
}
