% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{para_ll_fit_tele}
\alias{para_ll_fit_tele}
\title{Parallel Fit models on telemetry list}
\usage{
para_ll_fit_tele(tele_list)
}
\arguments{
\item{tele_list}{telemetry list}
}
\value{
list of model fitting results on each telemetry object, named by
telemetry object names
}
\description{
Parallel Fit models on telemetry list
}
\examples{
para_ll_fit_tele(buffalo)
}
