% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{para_ll_ud}
\alias{para_ll_ud}
\title{Parallel Calculate Occurrence From Telemetry And Model List}
\usage{
para_ll_ud(tele_model_list, fallback = FALSE)
}
\arguments{
\item{tele_model_list}{Aligned list of telemetry list and model list}

\item{fallback}{Use regular \code{lapply}. This is used to test if parallel caused
problems.}
}
\value{
occurrence results list
}
\description{
Parallel Calculate Occurrence From Telemetry And Model List
}
\examples{
para_ll_ud(align_list(buffalo, models_list))
}
