% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{para_ll_ud}
\alias{para_ll_ud}
\title{Title}
\usage{
para_ll_ud(ud_para_list)
}
\arguments{
\item{ud_para_list}{Aligned list of telemetry list and model list}
}
\value{
occurrence results list
}
\description{
Title
}
\examples{
para_ll_ud(align_list(buffalo, models_list))
}
