% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pick_m_tele}
\alias{pick_m_tele}
\title{Sample from telemetry object}
\usage{
pick_m_tele(tele, m)
}
\arguments{
\item{tele}{telemetry object}

\item{m}{sample size}
}
\value{
sampled telemetry object
}
\description{
Take even spaced m points. Rely on ctmm S3 method to treat telemetry object
as a data.frame.
}
\examples{
pick_m_tele(buffalo[[1]], 100)
}
