% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plot.R
\name{plot_loc}
\alias{plot_loc}
\title{Plot animal locations subset with all as background}
\usage{
plot_loc(loc_data_subset, loc_data = NULL, point_size = 0.1)
}
\arguments{
\item{loc_data_subset}{A subset of animals. A \code{data.table} subset taken from
full data set of all animals location. Note the \code{id} column is factor with
all animals identity as levels. So this subset still have all identities
information, which is needed to color each animal consistently.}

\item{loc_data}{The full data set of all animals location in merged
\code{data.table} format. Usually came from \link{combine}. If background is
not needed, it can be skipped and take the default \code{NULL} value. In this
case better use named parameter for \code{point_size} if needed, like
\code{plot_loc(loc_data_subset, point_size = 0.2)} instead of
\code{plot_loc(loc_data_subset, 0.2)} otherwise the second parameter will be
interpreted as \code{loc_data}.}

\item{point_size}{The size of point in plot. Denser plot may need smaller
point size.}
}
\value{
A ggplot object.
}
\description{
Draw a scatter plot of some animals' location with \code{ggplot2} (like plot 2
Overview in app). To add all animals as background, the full data set of
animals is also needed.
}
