% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_vario_hrange_occurrence.R
\name{plot_ud}
\alias{plot_ud}
\title{Plot a group of home ranges or occurrences}
\usage{
plot_ud(UD_list, level_vec = 0.95, color_vec = "blue",
  option = c("contour", "interval", "location"), columns = 2,
  cex = 0.65, tele_list = NULL)
}
\arguments{
\item{UD_list}{\code{ctmm} \code{UD} object list, which can be home range from
\code{\link[ctmm:akde]{ctmm::akde()}} or occurrences from \code{\link[ctmm:occurrence]{ctmm::occurrence()}}. The names of list
are needed for title of figures.}

\item{level_vec}{The vector of \code{level.UD} in \code{\link[ctmm:plot.telemetry]{ctmm::plot.telemetry()}}. To be
consistent with \code{ctmm} they are values 0 ~ 1 (for example 0.95). Note the
app UI take percentage values 0 ~ 100 (for example 95) for easier input.}

\item{color_vec}{The colors for contour, density and location points in each
plot by order. Single color is used for all plots.}

\item{option}{Whether to show contour, interval or location points. Note
interval only take effect when contour exsits.}

\item{columns}{The columns of the group plot layout.}

\item{cex}{The magnification factor of plot text and symbols. See \code{\link[=par]{par()}}.
You may need a smaller value if figure margins is too large}

\item{tele_list}{\code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} telemetry list. When provided, animal
locations are overlayed in plot. This should only be used for home range
since it can interfere with occurrence plot.}
}
\description{
This is a wrapper over \code{\link[ctmm:plot.telemetry]{ctmm::plot.telemetry()}} to make group plot easier.
}
