% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_vario_hrange_occurrence.R
\name{plot_vario}
\alias{plot_vario}
\title{Plot a group of empirical or modeled variograms.}
\usage{
plot_vario(vario_list, model_list = NULL, fraction = 0.5,
  relative_zoom = TRUE, cex = 0.65, model_color = "blue", columns = 2)
}
\arguments{
\item{vario_list}{list of \code{ctmm::variogram}. The \code{names} of list will be used
as figure title if in empirical variogram mode.}

\item{model_list}{\code{CTMM} models list. The \code{names} of list are needed for
figure titles. Draw modeled variogram if provided. The models list should
match \code{vario_list} in length and animal, so that \code{i}th model is for \code{i}th
variogram.
Possible values:
\itemize{
\item It can be skipped to use default value \code{NULL}, and only empirical
variograms are drawn. \code{vario_list} names will be used as figure title.
Better use named parameters for others in this case.
\item list of guesstimated models from \code{ctmm::ctmm.guess} on telemetry data,
overlay guesstimate variogram.
\item list of fitted models from \code{ctmm::ctmm.select} on \code{vario_list}, overlay
modeled variogram.
}}

\item{fraction}{Fraction of time-lag range, 0 ~ 1.}

\item{relative_zoom}{\itemize{
\item Relative mode zoom every plot by fraction of their own Time-lag range. The
X, Y axes are not synced.
\item Absolute mode operate on the max Time-lag range individual in group, and
all others scaled with same X, Y axes for easier comparison.
}}

\item{cex}{The magnification factor of plot text and symbols. See \link{par}.}

\item{model_color}{The color of model variogram}

\item{columns}{The columns of the group plot layout.}
}
\description{
This is a wrapper over \code{ctmm::plot.variogram} to make group plot easier.
}
