% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_map.R
\name{point_map}
\alias{point_map}
\title{Build point map from animal location data table}
\usage{
point_map(loc_data_subset)
}
\arguments{
\item{loc_data_subset}{\code{data.table} subset of full data set of animal locations from
\link{merge_tele}. The \code{id} column need to keep all animal names in levels to
maintain color consistency.}
}
\value{
A \code{Leaflet} map widget.
}
\description{
An interactive map will shown in RStudio Viewer pane when running in
interactive session. You can also further augment it with \code{leaflet}
functions, or save to a html with \code{htmlwidgets::saveWidget}.
}
