% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmm_data.R
\name{report}
\alias{report}
\title{Report data summary on telemetry object/list}
\usage{
report(tele_obj_list)
}
\arguments{
\item{tele_obj_list}{\code{\link[ctmm:as.telemetry]{ctmm::as.telemetry()}} telemetry object or list}
}
\value{
A summary \code{data.table}
}
\description{
Report data summary on telemetry object/list
}
