% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_parallel.R
\name{sample_tele}
\alias{sample_tele}
\title{Sample From Telemetry Object}
\usage{
sample_tele(tele, m)
}
\arguments{
\item{tele}{telemetry object}

\item{m}{sample size. \code{m} even spaced points are taken from data.}
}
\value{
sampled telemetry object
}
\description{
A sampled dataset can have models fitted much quicker. This is used to reduce
waiting time in developing code that involved time consuming modeling
processes. After code is tested and stablized, full size dataset can be used.
}
\examples{
library(ctmm)
data(buffalo)
sample_tele(buffalo[[1]], 100)
}
