% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_model_summary.R
\name{summary_model_fit}
\alias{summary_model_fit}
\title{Generate Formated Model Summary Table From Model Fit Results}
\usage{
summary_model_fit(model_fit_res)
}
\arguments{
\item{model_fit_res}{list of applying \code{ctmm::ctmm.select} on telemetry objects}
}
\value{
A \code{data.table} of model summary
}
\description{
Generate Formated Model Summary Table From Model Fit Results
}
\examples{
\dontrun{
 library(ctmm)
 library(ctmmweb)
 data("buffalo")
 data_sample <- sample_tele_list(buffalo, 100)
 model_fit_res <- par_fit_tele(data_sample)
 summary_model_fit(model_fit_res)
}
}
